change some stuff idk
add year in database

+++ app/controllers/Proyek.php
@@ -5,42 +5,58 @@ class Proyek extends Controller
     public function index()
     {
 
-        $data['judul'] = 'Daftar Proyek';
-        $data['proyek'] = $this->model('Proyek_model')->getAllProyek();
-        $this->view('templates/header', $data);
-        $this->view('proyek/index', $data);
-        $this->view('templates/footer');
+        if (isset($_SESSION['ptpwl_user_id'])) {
+            $data['judul'] = 'Daftar Proyek';
+            $data['proyek'] = $this->model('Proyek_model')->getAllProyek();
+            $this->view('templates/header', $data);
+            $this->view('proyek/index', $data);
+            $this->view('templates/footer');
+        } else {
+            header('location:' . BASEURL . 'login');
+        }
     }
     public function detail($id)
     {
-        $data['judul'] = 'Detail Proyek';
-        $data['proyek'] = $this->model('Proyek_model')->getProyekById($id);
-        $this->view('templates/header', $data);
-        $this->view('proyek/detail', $data);
-        $this->view('templates/footer');
+        if (isset($_SESSION['ptpwl_user_id'])) {
+            $data['judul'] = 'Detail Proyek';
+            $data['proyek'] = $this->model('Proyek_model')->getProyekById($id);
+            $this->view('templates/header', $data);
+            $this->view('proyek/detail', $data);
+            $this->view('templates/footer');
+        } else {
+            header('location:' . BASEURL . 'login');
+        }
     }
     public function tambah()
     {
-        if ($this->model('Proyek_model')->tambahDataProyek($_POST) > 0) {
-            Flasher::setFlash('berhasil', ' ditambahkan', 'success');
-            header('Location:' . BASEURL . 'proyek');
-            exit;
+        if (isset($_SESSION['ptpwl_user_id'])) {
+            if ($this->model('Proyek_model')->tambahDataProyek($_POST) > 0) {
+                Flasher::setFlash('berhasil', ' ditambahkan', 'success');
+                header('Location:' . BASEURL . 'proyek');
+                exit;
+            } else {
+                Flasher::setFlash('gagal', 'ditambahkan', 'danger');
+                header('Location:' . BASEURL . 'proyek');
+                exit;
+            }
         } else {
-            Flasher::setFlash('gagal', 'ditambahkan', 'danger');
-            header('Location:' . BASEURL . 'proyek');
-            exit;
+            header('location:' . BASEURL . 'login');
         }
     }
     public function hapus($id)
     {
-        if ($this->model('proyek_model')->hapusDataProyek($id) > 0) {
-            Flasher::setFlash('berhasil', ' dihapus', 'success');
-            header('Location:' . BASEURL . 'proyek');
-            exit;
+        if (isset($_SESSION['ptpwl_user_id'])) {
+            if ($this->model('proyek_model')->hapusDataProyek($id) > 0) {
+                Flasher::setFlash('berhasil', ' dihapus', 'success');
+                header('Location:' . BASEURL . 'proyek');
+                exit;
+            } else {
+                Flasher::setFlash('gagal', ' dihapus', 'danger');
+                header('Location:' . BASEURL . 'proyek');
+                exit;
+            }
         } else {
-            Flasher::setFlash('gagal', ' dihapus', 'danger');
-            header('Location:' . BASEURL . 'proyek');
-            exit;
+            header('location:' . BASEURL . 'login');
         }
     }
     public function getubah()
@@ -50,23 +66,30 @@ class Proyek extends Controller
 
     public function ubah()
     {
-        if ($this->model('Proyek_model')->ubahDataProyek($_POST) > 0) {
-            Flasher::setFlash('berhasil', ' diubah', 'success');
-            header('Location:' . BASEURL . 'proyek/detail/' . $_POST['id']);
-            exit;
+        if (isset($_SESSION['ptpwl_user_id'])) {
+            if ($this->model('Proyek_model')->ubahDataProyek($_POST) > 0) {
+                Flasher::setFlash('berhasil', ' diubah', 'success');
+                header('Location:' . BASEURL . 'proyek/detail/' . $_POST['id']);
+                exit;
+            } else {
+                Flasher::setFlash('gagal', ' diubah', 'danger');
+                header('Location:' . BASEURL . 'proyek/detail/' . $_POST['id']);
+                exit;
+            }
         } else {
-            Flasher::setFlash('gagal', ' diubah', 'danger');
-            header('Location:' . BASEURL . 'proyek/detail/' . $_POST['id']);
-            exit;
+            header('location:' . BASEURL . 'login');
         }
     }
     public function cari()
     {
-
-        $data['judul'] = 'Daftar Proyek';
-        $data['proyek'] = $this->model('Proyek_model')->cariDataProyek();
-        $this->view('templates/header', $data);
-        $this->view('proyek/index', $data);
-        $this->view('templates/footer');
+        if (isset($_SESSION['ptpwl_user_id'])) {
+            $data['judul'] = 'Daftar Proyek';
+            $data['proyek'] = $this->model('Proyek_model')->cariDataProyek();
+            $this->view('templates/header', $data);
+            $this->view('proyek/index', $data);
+            $this->view('templates/footer');
+        } else {
+            header('location:' . BASEURL . 'login');
+        }
     }
 }
